feat: retry installer downloads and rename to Agentic Tools (#5247)
Retry urlopen/Maven paths on transient SSL/URL/timeout, keep a persistent logo and overall progress, and make install-shaft-mcp* deprecation shims.
